+++ src/main/kotlin/dev/slint/ideaplugin/ide/settings/SlintStateData.kt
@@ -1,5 +1,0 @@
-package dev.slint.ideaplugin.ide.settings
-
-data class SlintStateData(
-    var appSettings: SlintLspSettings = SlintLspSettings(),
-)